@@ -1,13 +1,20 @@
-import { DisplayProcessor, SpecReporter, StacktraceOption } from 'jasmine-spec-reporter'
-import SuiteInfo = jasmine.SuiteInfo
+import {
+  DisplayProcessor,
+  SpecReporter,
+  StacktraceOption,
+} from 'jasmine-spec-reporter';
+import SuiteInfo = jasmine.SuiteInfo;
 
 class CustomProcessor extends DisplayProcessor {
-  public displayJasmineStarted(info: SuiteInfo, log: string): string {
-    return `TypeScript ${log}`
+  public displayJasmineStarted(
+    info: SuiteInfo,
+    log: string
+  ): string {
+    return `TypeScript ${log}`;
   }
 }
 
-jasmine.getEnv().clearReporters()
+jasmine.getEnv().clearReporters();
 jasmine.getEnv().addReporter(
   new SpecReporter({
     spec: {
@@ -15,4 +22,4 @@ jasmine.getEnv().addReporter(
     },
     customProcessors: [CustomProcessor],
   })
-)
+);
